requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
idna==3.11
pillow==12.2.0
psycopg==3.3.3
psycopg-binary==3.3.3
reportlab==4.4.10
requests==2.33.1
sqlparse==0.5.5
stripe==15.0.1
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
